ubuntu/streamlit_app/streamlit_gdrive_app/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
